implemented payment feature with stripe

+++ client/src/app/checkout/checkout-payment/checkout-payment.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { AfterViewInit, Component, ElementRef, Input, OnDestroy, OnInit, ViewChild } from '@angular/core';
 import { FormGroup } from '@angular/forms';
 import { NavigationExtras, Router } from '@angular/router';
 import { ToastrService } from 'ngx-toastr';
@@ -7,36 +7,99 @@ import { IBasket } from 'src/app/shared/models/basket';
 import { IOrderToCreate } from 'src/app/shared/models/order';
 import { CheckoutService } from '../checkout.service';
 
+declare var Stripe: any;
+
 @Component({
   selector: 'app-checkout-payment',
   templateUrl: './checkout-payment.component.html',
   styleUrls: ['./checkout-payment.component.scss']
 })
-export class CheckoutPaymentComponent implements OnInit {
+export class CheckoutPaymentComponent implements AfterViewInit, OnDestroy {
   @Input() checkoutForm: FormGroup = null!;
+  @ViewChild('cardNumber', {static: true}) cardNumberElement: ElementRef = new ElementRef(null);
+  @ViewChild('cardExpiry', {static: true}) cardExpiryElement: ElementRef = new ElementRef(null);
+  @ViewChild('cardCvc', {static: true}) cardCvcElement: ElementRef = new ElementRef(null);
+  stripe: any;
+  cardNumber: any;
+  cardExpiry: any;
+  cardCvc: any;
+  cardErrors: any;
+  cardHandler: any = this.onChange.bind(this);
+  loading = false;
 
   constructor(
     private basketService: BasketService, 
     private checkoutService: CheckoutService,
     private toastr: ToastrService,
     private router: Router) { }
-
-  ngOnInit(): void {
+  
+  ngOnDestroy(): void {
+    this.cardNumber.destroy();
+    this.cardExpiry.destroy();
+    this.cardCvc.destroy();
   }
 
-  submitOrder() {
+  ngAfterViewInit(): void {
+    this.stripe = Stripe('pk_test_51LzFINGC9GzJMueBQa8XwTe09K3rPMR6oH2lmQxeFzocpgY6bJrP0zvl5gBKkmL8jEjxUWt8iEehCHSl9PJq0gxZ00NkrsBoVr');
+    const elements = this.stripe.elements();
+
+    this.cardNumber = elements.create('cardNumber');
+    this.cardNumber.mount(this.cardNumberElement.nativeElement);
+    this.cardNumber.addEventListener('change', this.cardHandler);
+
+    this.cardExpiry = elements.create('cardExpiry');
+    this.cardExpiry.mount(this.cardExpiryElement.nativeElement);
+    this.cardExpiry.addEventListener('change', this.cardHandler);
+    
+    this.cardCvc = elements.create('cardCvc');
+    this.cardCvc.mount(this.cardCvcElement.nativeElement);
+    this.cardCvc.addEventListener('change', this.cardHandler);
+    
+  }
+  onChange({error}:any) {
+    if(error) {
+      this.cardErrors = error.message;
+    } else {
+      this.cardErrors = null;
+    }
+  }
+
+  async submitOrder() {
+    this.loading = true;
     const basket = this.basketService.getCurrentBasketValue();
-    const orderToCreate = this.getOrderToCreate(basket!);
-    this.checkoutService.createOrder(orderToCreate).subscribe({
-      next: (order) => {
-        this.toastr.success('Order created successfully');
+    try {
+      const createdOrder = await this.createOrder(basket!);
+      const paymentResult = await this.confirmPaymentWithStripe(basket!);
+      if (paymentResult.paymentIntent) {
         this.basketService.deleteLocalBasket(basket!.id);
-        const navigationExtras: NavigationExtras = {state:order}
-        this.router.navigate(['/checkout/success'], navigationExtras);
-      },
-      error: (error) => this.toastr.error(error.message)
-    })
+        const navigationExtras: NavigationExtras = { state: createdOrder };
+        this.router.navigate(['checkout/success'], navigationExtras);
+      } else {
+        this.toastr.error(paymentResult.error.message);
+      }
+      this.loading = false;
+    } catch (error) {
+      console.log(error);
+      this.loading = false;
+    }
   }
+
+  private async confirmPaymentWithStripe(basket: IBasket) {
+    return this.stripe.confirmCardPayment(basket.clientSecret, {
+      payment_method: {
+        card: this.cardNumber,
+        billing_details: {
+          name: this.checkoutForm.get('paymentForm')?.get('nameOnCard')?.value
+        }
+      }
+    });
+  }
+
+  private async createOrder(basket: IBasket) {
+    const orderToCreate = this.getOrderToCreate(basket);
+    return this.checkoutService.createOrder(orderToCreate).toPromise();
+  }
+
   private getOrderToCreate(basket: IBasket) : IOrderToCreate {
     return{
       basketId: basket.id,
